Add support for file level instrumentation

+++ test/tryLLVM/Basic/instrument/mop.c
@@ -3,6 +3,6 @@
 
 //#include "prthread.h"
 
-void mopInstrument(int address, int typeSize, char* type, char* debugLoc, char *fName){
+void mopInstrument(long long address, int typeSize, char* type, char* debugLoc, char *fName){
     printf("access (??, %p) of type %s (typesize %d) at %s\n", address, type, typeSize, debugLoc);
 }
